% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioDataScience1-package.R
\docType{package}
\name{BioDataScience1-package}
\alias{BioDataScience1-package}
\title{A series of learnr tutorials and Shiny applications for BioDataScience1}
\description{
Run the tutorials using\code{\link[=run]{run()}}.
}
