% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Configure the R environment to acces my MongoDB database}
\usage{
config(debug = Sys.getenv("LEARNDOWN_DEBUG", 0) != 0)
}
\arguments{
\item{debug}{Do we issue debugging messages (by default, yes if the
environment variables \code{LEARNDOWN_DEBUG} is not \code{0}).}
}
\value{
\code{TRUE} or \code{FALSE} invisibly, if it succeeds or not (with database
access test).
}
\description{
Environment variables are added with infos about my database.
}
